server/exporter/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
pyarrow==15.0.0
schedule==1.2.1
